requirements.txt
flask==3.0.0
scikit-learn==1.3.2
numpy==1.24.3
nltk==3.8.1
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
unidecode==1.3.7
openai==1.3.0
python-dotenv==1.0.0
